Distributed functions over more files
Distributed functions over more files, for better readability.
