+++ src/CMakeLists.txt
@@ -1,5 +1,5 @@
 # Add package_distribution as a library
-add_library(package_functions package_distribution.c delivery_algorithm.c delivery_sim.c resource_calculation.c create_routes.c)
+add_library(package_functions package_distribution.c delivery_algorithm.c delivery_sim.c resource_calculation.c create_routes.c bst_functions.c astar_helper_functions.c)
 target_include_directories(package_functions PUBLIC .)
 
 # Add executable file
