+++ src/delivery_algorithm.h
@@ -1,16 +1,3 @@
 #include "package_distribution.h"
 
-void add_node_to_tree(node_t *new_node, tree_t *tree);
-void add_node_to_tree_rec(node_t *new_node, node_t *tree_node);
-node_t *find_lowest_f_in_tree(node_t *tree_root);
-int check_in_tree(node_t *node, node_t *tree_root);
-int *reconstruct_path(node_t *node, int nodes_amount);
-void reconstruct_path_rec(node_t *node_parent, int *list, int *count);
-
-double heuristic(node_t current_node, node_t current_node_neighbour);
-
 void a_star(graph_t *graph, a_star_matrix_t *a_star_matrix, node_t start_node, node_t end_node);
-
-void remove_node_from_tree(node_t *node, tree_t *root);
-node_t *find_successor(node_t *node);
-node_t *find_parent(node_t *node, tree_t *tree);
